Bookworm: minor

+++ modules/prebookworm/templates/update_source_debian.sh
@@ -11,6 +11,8 @@ deb http://deb.debian.org/debian bookworm-updates main
 deb-src http://deb.debian.org/debian bookworm-updates main" > /etc/apt/sources.list
 
 #delete backports
-rm /etc/apt/sources.list.d/backports.list
+if [ -f rm /etc/apt/sources.list.d/backports.list ]; then
+  rm /etc/apt/sources.list.d/backports.list
+fi
 
 cat /etc/apt/sources.list
